Begin stripe integration, Openlaw error checks
Now we check the status of the openlaw JWT request and show the user an
error if we get anything fishy, including high status code or empty jwt.

Added the stripe library to the index and heavenly interface via the
next/head library. I think I changed the title of the pages though as
well, sorry Tristan! No more sparkles for you!