Stripe checkout and payment choice modal
When the user clicks the create button they are shown a payment choice
modal that shows them the price and options to pay in fiat and crypto.
If they click a payment in fiat or crypto, their openlaw doc is sent to
our help email and they are shown the crypto address or taken to the
stripe checkout window.

The stripe checkout flow has been tested for proper callback but the
page that the user should land on when they complete their payment or
cancel it hasn't been tested yet. In addition, I'm not sure yet how to
keep track of the users' openlaw form details if they cancel a stripe
payment.