Added support for Openlaw errors w/ contracts
Now if the contract fails to upload we upload a draft and send it to
ourselves. TODO: implenment two server-side APIs documented in API.ts.

If the contract uploads but fails to send we rename it to Error -
discard and send us the draft and an email about what happened.

See API.ts for the two APIs we need implemented.